Rename NametagEntity to LabelEntity for clarity
